We have a horribly inefficient game-mode :D
Seriously, the performance/artifacts are horrible. I'll have to spend a
moment gathering data and optimizing the hell out of this. In the worst
case I'd stop all the concurrent shit when running the game, kill the
editor, and make it possible to swap between the editor and the game
without losing most of the data.

Plus: I'll have to implement a thing where running the game doesn't
screw with editor's initial values